data_processer/requirements.txt
pika==1.3.1
python-dotenv==0.21.0
redis==4.4.2
h3==3.7.6
pendulum==2.1.2
pygelf==0.4.2